move code to physical_plan_builder.go
